backend/requirements.txt
asgiref==3.5.2
boto3==1.26.112
botocore==1.29.112
Django==4.1.4
django-cors-headers==3.13.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
jmespath==1.0.1
jsonfield==3.1.0
mysql-connector-python==8.0.33
mysqlclient==2.1.1
Pillow==9.3.0
protobuf==3.20.3
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
whitenoise==6.3.0
